Android: Call nsd.unregisterService() synchronously when going to sleep
If Android is on it's way to suspend when stopServicePublish() is called,
we need to call nsd.unregisterService() synchronously to force it to run
before the device goes to sleep.  If instead it is scheduled to run in the
Android thread, it will not run until the device is woken back up.