Fix Android api 30 issue
Google disabled bind() to a netlink socket in api 30.  This change disables
the netlink socket used in avahi-core on Android which means there is no
async network interface info.  When the apps starts, a list of interfaces
is gotten with getifaddrs() and avahi uses this list to monitor.  Avahi
does not monitor while the app sleeps and deletes the monitor.