Bonjour - destroy resolvers by resolvers list, not service list
If stopBrowser is called after a resolve starts, but before the resolve
finished, the service is never added to the services list.  This leaves the
resolver active and it fires when resolving completes after stopBrowser()
has destroyed a bunch of objects.  Fix is to destroy resolves according to
resolvers list and destroy services according to services list.